// powerapps-project-planner-20260612/src/generated/models/Vibe_projecttemplatesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Vibe_projecttemplatesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Vibe_projecttemplatesstatecode = keyof typeof Vibe_projecttemplatesstatecode;
export const Vibe_projecttemplatesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Vibe_projecttemplatesstatuscode = keyof typeof Vibe_projecttemplatesstatuscode;

export interface Vibe_projecttemplatesBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Vibe_projecttemplatesstatecode;
  statuscode?: Vibe_projecttemplatesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
  vibe_description?: string;
  vibe_isactive?: boolean;
  vibe_name?: string;
  vibe_projecttemplateid: string;
}

export interface Vibe_projecttemplates extends Vibe_projecttemplatesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  vibe_isactivename?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}