// powerapps-project-planner-20260612/src/generated/models/Vibe_phasetemplatesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Vibe_phasetemplatesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Vibe_phasetemplatesstatecode = keyof typeof Vibe_phasetemplatesstatecode;
export const Vibe_phasetemplatesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Vibe_phasetemplatesstatuscode = keyof typeof Vibe_phasetemplatesstatuscode;

export interface Vibe_phasetemplatesBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Vibe_phasetemplatesstatecode;
  statuscode?: Vibe_phasetemplatesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
  vibe_defaultdurationdays?: number;
  vibe_name?: string;
  vibe_order?: number;
  vibe_phasetemplateid: string;
  "vibe_projecttemplateid@odata.bind"?: string;
}

export interface Vibe_phasetemplates extends Vibe_phasetemplatesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  vibe_projecttemplateidname?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  vibe_projecttemplateid?: object;
  _vibe_projecttemplateid_value?: string;
}