// powerapps-project-planner-20260612/src/generated/models/Vibe_tasktemplatesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Vibe_tasktemplatesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Vibe_tasktemplatesstatecode = keyof typeof Vibe_tasktemplatesstatecode;
export const Vibe_tasktemplatesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Vibe_tasktemplatesstatuscode = keyof typeof Vibe_tasktemplatesstatuscode;
export const Vibe_tasktemplatesvibe_defaultpriority = {
  100000000: 'LOW',
  100000001: 'MEDIUM',
  100000002: 'HIGH',
  100000003: 'CRITICAL'
} as const;
export type Vibe_tasktemplatesvibe_defaultpriority = keyof typeof Vibe_tasktemplatesvibe_defaultpriority;

export interface Vibe_tasktemplatesBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Vibe_tasktemplatesstatecode;
  statuscode?: Vibe_tasktemplatesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
  vibe_defaultestimatedhours?: number;
  vibe_defaultpriority?: Vibe_tasktemplatesvibe_defaultpriority;
  "vibe_defaultroleid@odata.bind"?: string;
  vibe_name: string;
  vibe_order?: number;
  "vibe_phasetemplateid@odata.bind"?: string;
  vibe_tasktemplateid: string;
}

export interface Vibe_tasktemplates extends Vibe_tasktemplatesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  vibe_defaultpriorityname?: string;
  vibe_defaultroleidname?: string;
  vibe_phasetemplateidname?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  vibe_defaultroleid?: object;
  _vibe_defaultroleid_value?: string;
  vibe_phasetemplateid?: object;
  _vibe_phasetemplateid_value?: string;
}